requirements.txt
PyYAML==6.0.1
beautifulsoup4==4.12.2
canvasapi==3.2.0
jsonpickle==3.0.2
requests==2.31.0
tqdm==4.66.1
python-dateutil==-2.8.2